Move GameState to own file
